+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+-   Passing Kedro environment and pipeline name in Vertex nodes
+-   Setting artifact type based on catalog layer in Vertex pipeline
+-   Added `pipeline` param to `schedule` in Vertex
+
 ## [0.4.0] - 2021-08-11
 
 -   Support for kedro-mlflow>=0.7